refactor: Split up constants into individual files
