+++ libs/common/constants/src/index.ts
@@ -1,1 +1,8 @@
-export * from './lib/common-constants'
+export * from './lib/alts'
+export * from './lib/factions'
+export * from './lib/bases'
+export * from './lib/weapons'
+export * from './lib/vehicles'
+export * from './lib/continents'
+export * from './lib/servers'
+export * from './lib/infantry'
