+++ libs/common/constants/src/lib/alts.ts
@@ -1,0 +1,12 @@
+export const ps2gg = {
+  altsUrl: 'http://alts:3003',
+}
+
+export const roleThresholds = {
+  pilot: {
+    esf: {
+      playTimePercent: 0.2,
+      kills: 3000,
+    },
+  },
+}
