+++ libs/common/constants/src/lib/continents.ts
@@ -1,0 +1,7 @@
+export const continents: { [key: string]: string } = {
+  '2': 'Indar',
+  '4': 'Hossin',
+  '6': 'Amerish',
+  '8': 'Esamir',
+  '344': 'Oshur',
+}
