+++ libs/common/constants/src/lib/factions.ts
@@ -1,0 +1,6 @@
+export const factions: { [key: string]: string } = {
+  '1': 'VS',
+  '2': 'NC',
+  '3': 'TR',
+  '4': 'NS',
+}
