+++ libs/common/constants/src/lib/infantry.ts
@@ -1,0 +1,26 @@
+export const infantry: { [key: string]: string } = {
+  '1': 'Infiltrator',
+  '3': 'Light Assault',
+  '4': 'Medic',
+  '5': 'Engineer',
+  '6': 'Heavy Assault',
+  '7': 'MAX',
+  '8': 'Infiltrator',
+  '10': 'Light Assault',
+  '11': 'Medic',
+  '12': 'Engineer',
+  '13': 'Heavy Assault',
+  '14': 'MAX',
+  '15': 'Infiltrator',
+  '17': 'Light Assault',
+  '18': 'Medic',
+  '19': 'Engineer',
+  '20': 'Heavy Assault',
+  '21': 'MAX',
+  '28': 'Infiltrator',
+  '29': 'Light Assault',
+  '30': 'Medic',
+  '31': 'Engineer',
+  '32': 'Heavy Assault',
+  '45': 'MAX',
+}
