+++ libs/common/constants/src/lib/servers.ts
@@ -1,0 +1,8 @@
+export const servers: { [key: string]: string } = {
+  '1': 'Connery',
+  '10': 'Miller',
+  '13': 'Cobalt',
+  '17': 'Emerald',
+  '19': 'Jaeger',
+  '40': 'SolTech',
+}
