+++ libs/common/constants/src/lib/vehicles.ts
@@ -1,0 +1,29 @@
+export const vehicles: { [key: string]: string } = {
+  '1': 'Flash',
+  '2': 'Sunderer',
+  '3': 'Lightning',
+  '4': 'Magrider',
+  '5': 'Vanguard',
+  '6': 'Prowler',
+  '7': 'Scythe',
+  '8': 'Reaver',
+  '9': 'Mosquito',
+  '10': 'Liberator',
+  '11': 'Galaxy',
+  '12': 'Harasser',
+  '14': 'Valkyrie',
+  '15': 'Ant',
+  '2019': 'Bastion',
+}
+
+export const esf: { [key: string]: string } = {
+  '7': 'Scythe',
+  '8': 'Reaver',
+  '9': 'Mosquito',
+}
+
+export const mbt: { [key: string]: string } = {
+  '4': 'Magrider',
+  '5': 'Vanguard',
+  '6': 'Prowler',
+}
