+++ libs/common/constants/src/lib/weapons.ts
@@ -1,0 +1,35 @@
+export const a2aWeapons: { [key: string]: string } = {
+  '4900': 'M18 Needler',
+  '4302': 'Saron Laser Cannon',
+  '4600': 'M20 Mustang',
+
+  '4905': 'Tomcat A2AM Pods',
+  '4300': 'Photon A2A Missile Pods',
+  '4602': 'Tomcat A2AM Pods',
+
+  '4911': 'M18 Rotary',
+  '4304': 'Maelstrom Turbo Laser',
+  '4604': 'Vortek Rotary',
+
+  '5050': 'M18 Locust',
+  '4445': 'Antares LC',
+  '4745': 'M20 Kestrel',
+
+  '5052': 'Coyote Missiles',
+  '4447': 'Coyote Missiles',
+  '4747': 'Coyote Missiles',
+}
+
+export const a2gWeapons: { [key: string]: string } = {
+  '4903': 'Hellfire Rocket Pods',
+  '4301': 'Dual Photon Pods',
+  '4601': 'Breaker Rocket Pods',
+
+  '4906': 'M14 Banshee',
+  '4305': 'Light PPA',
+  '4605': 'M30 Mustang AH',
+
+  '5051': 'Hornet Missiles',
+  '4446': 'Hornet Missiles',
+  '4746': 'Hornet Missiles',
+}
